+++ detox/test/types/detox-jest-setup-tests.ts
@@ -2,9 +2,9 @@ declare var beforeAll: (callback: () => void) => void;
 declare var beforeEach: (callback: () => void) => void;
 declare var afterAll: (callback: () => void) => void;
 
-import detox = require("detox");
-import adapter = require("detox/runners/jest/adapter");
-import specReporter = require("detox/runners/jest/specReporter");
+import * as detox from 'detox';
+import * as adapter from 'detox/runners/jest/adapter';
+import * as specReporter from 'detox/runners/jest/specReporter';
 
 // Normally the Detox configuration from the project's package.json like so:
 // const config = require("./package.json").detox;
